@@ -2,30 +2,30 @@
   "name": "MeepleBnB",
   "private": true,
   "dependencies": {
-    "@babel/core": "^7.1.5",
-    "@babel/preset-env": "^7.1.5",
-    "@babel/preset-react": "^7.0.0",
-    "@emotion/core": "^10.0.17",
-    "@emotion/styled": "^10.0.17",
-    "babel-loader": "^8.0.4",
+    "@babel/core": "^7.11.6",
+    "@babel/preset-env": "^7.11.5",
+    "@babel/preset-react": "^7.10.4",
+    "@emotion/core": "^10.0.35",
+    "@emotion/styled": "^10.0.27",
+    "babel-loader": "^8.1.0",
     "css-loader": "^1.0.1",
-    "jquery": "^3.3.1",
-    "moment": "^2.22.2",
-    "react": "^16.6.3",
+    "jquery": "^3.5.1",
+    "moment": "^2.27.0",
+    "react": "^16.8.0",
     "react-addons-shallow-compare": "^15.6.2",
-    "react-dates": "^18.2.2",
-    "react-dom": "^16.6.3",
+    "react-dates": "^18.5.0",
+    "react-dom": "^16.8.0",
     "react-onclickoutside": "^6.9.0",
     "react-rating": "^2.0.5",
-    "react-redux": "^5.1.1",
+    "react-redux": "^5.1.2",
     "react-router-dom": "^4.3.1",
-    "react-select": "^3.0.4",
-    "react-with-direction": "^1.3.0",
-    "redux": "^4.0.1",
+    "react-select": "^3.1.0",
+    "react-with-direction": "^1.3.1",
+    "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
-    "webpack": "^4.25.1",
-    "webpack-cli": "^3.1.2"
+    "webpack": "^4.44.1",
+    "webpack-cli": "^3.3.12"
   },
   "description": "This README would normally document whatever steps are necessary to get the application up and running.",
   "version": "1.0.0",
